Add coremark benchmarking app
This is a port of https://github.com/eembc/coremark.

The app runs a suite of tests and delivers some performance measure at
the end of the test. The test must run for 10s so the COREMARK_ITERATION
might need to be tweaked according to the speed of the MCU running it.
A succesfull output will be something like:

2K performance run parameters for coremark.
CoreMark Size    : 666
Total ticks      : 15538
Total time (secs): 15
Iterations/Sec   : 133
Iterations       : 2000
Compiler version : GCC9.1.0
Compiler flags   :
Memory location  : STACK
seedcrc          : 0xe9f5
[0]crclist       : 0xe714
[0]crcmatrix     : 0x1fd7
[0]crcstate      : 0x8e3a
[0]crcfinal      : 0x4983
Correct operation validated. See README.md for run and reporting rules.

NOTE: since this is CPU intensive and running on the main thread, it
might trigger the watchdog if COREMARK_ITERATION is a large number.