hw/mcu/dialog: SNC (#1868)
* hw/mcu/dialog: SNC

Sensor node controller API. This commit allows users to create
SNC programs using the macros in da1469x_snc.h and run them
either under software control or use the PDC to start them. Note
that PDC control of the SNC is not contained here; the PDC api
already contain all that is needed to have the PDC start the
SNC.

Added in this commit is a test program (hosted) that illustrates
how a program can be created and run under software control. The
code for the test program is in hw/mcu/dialong/da1469x/hosttest.
This directory is also intended for other host tests and not
just the SNC test.

* hw/mcu/dialog: SNC

This commit adds interrupt registering and handling code for the SNC. Adds additional test cases to test the interrupt handling API. Addressed PR review comments as well.

* hw/mcu/dialog/da1469x/hosttest: Add license file to syscfg.yml