libc/baselibc: Fix memset on ARM
If buffer is unaligned and fits in single word, the number of unaligned
bytes at the end of buffer is calculated as a negative number. This
causes memset to overwrite few bytes before actual buffer.

To fix this we just need to check if calculated value is negative and
use the other routine to fill unaligned bytes.