sensors/bmp388: Fix init_intpin
Device has only one interrupt pin. There is no need for searching
table for pin configuration.
Search code that is shared with other drivers would not work
as intended since host_pin is uint8_t so it is never negative.
Now if BMP388_INT_ENABLE is on pin configuration is used
as it is.