Remove `LOG_MODULE_PERUSER`
Now that module IDs can be remapped via syscfg, there is no need to
partition modules into two groups ("system" and "user").

A consequence of this commit is that the array of module names no longer
starts at index 64.  To prevent applications from failing to register
module names, this commit introduces a data structure to represent a
registered module.  The structure contains the ID of the registered
module rather than using the array index.