riscv: Debugger detection
This is workaround for debugger detection and can be
used till better way is found.

Debugger sets magic value to one of hardware registers that
is not used for anything else.
hal_debugger_connected() checks this register for magic value
set by debugger script.
Debugger disconnection will not be detected.