sensors/lis2dh12: Allow disabling one interface
In real life scenario accelerometer will be soldered to the board
and only one interface will be used to access it.
Current code always has built-in function for I2C and SPI due to
runtime checking of interface type.
This change adds syscfg values to enable interface in build
(default both are enabled for old time sake).

This allows to reduce image size.