sys/console: insert_char slightly updated
Function insert_char was a little bit incoherent.
Third argument 'end' in all the calls was global variable
'trailing_chars' (former 'end').
Function was modifying global variable 'cur' which is closely
related to 'trailing_chars'.

Third argument was removed from this static function and now it is
using global 'trailing_chars' same way as it is using 'cur'.

Function was also sending terminal sequences for saving and restoring
cursor position but it was not checking echo variable as it should.

Now restoring cursor position is done by cursor_backward() which
will be called only if echo is enabled.
Function cursor_backward() will also not send any sequence to terminal
if cursor is already at the end of the line.