hw/mcu/dialog: Allow for internal pullup on RX pin
Refactored the code to add the rx_pullup flag to the config
as this is really a BSP item. Added a pointer to the config
to the driver to get the pin and also polled UART busy as
opposed to checking it once as it may take a bit of time
for the line to be pulled up.