sys/console: Add software cursor in prompt area
For massive log output it can be difficult to see
cursor position in prompt area.
In that case software rendered cursor is useful.
This adds such functionality.
Cursor is rendered inverse by default which is suitable
for black and white only terminal.
For color terminal escape sequence can be set to make
software cursor matching with hardware one.