stm32: updates to flash drivers
This applies a few refactorings that were suggested in a previous PR.

Remove syscfg parameters to specify if a flash has single sized sectors
and the size of those sectors. Families that have single sized sectors
export a `FLASH_PAGE_SIZE` definition, which can also be used to
determine the size of the pages.

A few devices have page sizes that are too small and would use too much
RAM, so for those devices emulate sectors that are slighly larger.

Underscore prefixes where removed from #defines.