sys/flash_map: Keep non-overlapping default areas
Prior to this commit, a device with a flash map in its manufacturing
meta region (MMR) would completely ignore the default flash map
hardcoded in a firmware image.  The rationale was that changes to the
default flash map should not cause devices in the field to lose data
present in the affected flash areas.

This is reasonable, but there is one issue - if we add new flash areas
to unused portions of flash, devices in the field will not be able to
use them since they would not be present in the devices' flash maps.

This commit allows a device to use new flash areas hardcoded in the
image.  The device will incorporate such a flash area into its flash map
if it meets both of these criteria:

1. The area has a unique ID
2. The area does not overlap with any other flash areas in the MMR flash
   map.

The first criterion means that if a firmware image rearranges or resizes
some existing flash areas, these changes will *not* be incorporated into
the device's flash map.  These areas are not new, they are just
modified, so their IDs are already present in the device's MMR flash
map.

The second criterion means that the device will only incorporate a new
flash area into its flash map if the area resides entirely in unused
flash.  If the area overlaps any other areas, it is ignored.