sys/console: Add console history persistence
With this change console history can be stored in log.
If CONSOLE_HISTORY_LOG_NAME is set to name of log it
will be used to store console history.

If log name is NOT specified but there is dedicated flash
area FLASH_AREA_CONSOLE_HISTORY and FCB2 is used, this
flash area will be used.