kernel/os: Fix sanity interval handling
Previous check causes idle task to skip sanity check if wake up occured
at the same tick when sanity should happen. This should not be a problem
because iticks will be 0 and sanity check will be performed after wakeup
on next tick. The only side effect should be wake up on 2 consecutive
ticks. However...

It is (unlikely but) possible that (at least) one tick occurs between
check and entering critical section. If this happens on a tick where
sanity check is skipped, time to next sanity is calculated as a negative
value instead of 0 so depending on time to next sched/callout we may not
wake up for sanity and tickle watchdog on time.