sys/config: Expose `conf_lock` and conf_unlock`
Prior to this commit, these two functions were private to the config
package.  This commit moves their prototypes to the public `config.h`
header.

If another task tries to read or write a setting while the config
package is locked, the operation will block until the package is
unlocked.  The locking task is permitted to read and write settings as
usual.

Typical usage of the config API does not require manual locking.  This
functionality is only needed for unusual use cases such as temporarily
changing the destination medium for an isolated series of writes.