da1469x: Add Encrypted Flash Driver
Add Encrypted Flash Driver that makes use of the AES encrypt engine
on the SoC. The driver supports CTR mode for 256/192/128 bits.

The AES Key which is stored in the OTP area is securely DMA transferred
into the AES engine during the encrypt operation.

The OTP area has slots (0-7) that could be used for storing the keys.
The slot number is specified with a MYNEWT configuration parameter.

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>