hw/mcu/nordic: Fix hal_timer firing too early
In hal_timer there is a workaround for an RTC compare issue which in
some cases fires timer callback too early. The problem with this code
is that timer value read in a callback will be *before* what was set
on timer and this is not what callback expects.

To fix this, in case we detect that new compare is too close to current
timer value we should not trigger timer too early but instead enable
TICK interrupt which occurs on every counter change and disable it once
condition for a workaround is no longer valid. This can cause an extra
interrupt on a tick with no timers being handled, but this is ok.