hw/drivers/crypto: Add crypto_in_use flag. (#2134)
* hw/drivers/crypto: Add crypto_in_use flag.

This commit adds a flag to the crypto driver called "in_use".
This flag, and associated API, is intended for use by ISR's to
perform a quick check to see if an underlying crypto hw peripheral
is being used so that the HW peripheral can be used in the ISR
or the use of an associated piece may impact an ongoing crypto
operation.

Currently, this is only implemented for the nrf52. Some changes
were made to the nrf52 crypto code. Some reasoning behind some
of the changes: 1) do not think a counter to bail out of the
loop waiting for the crypto to end or an error to occur is needed.
Either one of the two will occur. 2) Not quite sure if it is
possible but if both EVENTS_ENDECB and EVENTS_ERRORECB can be set
at the same time, the previous could assume success when there was
an error. 3) If EVENTS_ERRORECB is generated we simply try the
operation again.

Note that a somewhat un-related change was made to the encrypted
flash ef_crypto driver to add a restriction such that CRYPTO is
required to be defined when using this package.

* hw/drivers/crypto: Use bool for in use function

Per review comments, use bool instead of int for crypto_in_use
API.