Fix stack sizes and missing stack alignments
This fixes issues with memory exhaustion running util/rwlock/selftest
and kernel/os/selftest due to not using the proper `OS_STACK_ALIGN`
macros. Because of the lack of control over memory usage when running
on the sim, this macro multiplies the stack size by 16, to guarantee
that spare memory is available.

Signed-off-by: Fabio Utzig <utzig@apache.org>