hw/hal: Add generic API to retrieve provisioned data
Currently we only have dedicated API to retrieve HW ID from BSP.
However, there are more data that can be provisioned to device and it
would be nice to be able to retrieve it in a consistent way regardless
of BSP used.

This patch adds generic API to retrieve provisioned data. Note that it
does not specify how these data should be provisioned to device, it's
just about reading them.

There are few predefined data identifiers and this list can be extended
as there are more data that are usable by various packages. Also there
is separate range for user-defined identifiers.

In general provisioned data are BSP-specific but application can provide
own callback to override some data or support custom data identifiers
via the same common API.