hw/mcu/dialog: Split DCDC startup into 2 phases
Currently DCDC configuration is hardcoded and it's not really possible
to change it in BSP. This patch makes it easier by splitting DCDC
startup into 2 phases: initialization and enabling.

Initialization is done in SystemInit and it only configures registers.
DCDC is then enabled in hal_system_init. This allows to inject extra
configuration in BSP startup code by modifying DCDC registers between
both calls and proper configuration will be retained on wakeup as
expected.