da1469x: otp and serial script enhancements (#2262)
* hw/bsp/da1469x: da1469x_serial: Make the reset script optional

Make the reset script an optional parameter. This gives the
user an option to manually reset the board to make it enter
the serial load mode or use JLink or OpenOCD based reset script.

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>

* hw/bsp/da1469x: otp_tool: provide common function for serial

Provide common method for serial port opening opens the port, flushes
and returns port handle. This eleminates repeated code.

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>

* hw/bsp/da1469x: da1469x_serial: drain the port after open

Drain the port after opening to ensure the
buffers are empty.

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>