encoding/base64: Support decode in chunks
Add support for chunked decoding.  This is done with a new struct which
remembers the partial token it read from the previous (incomplete)
chunk.

This also fixes two bugs related to base64 decoding:

1. Never read past the end of the source string (even if its length is
not a multiple of four).

2. Return error (-1) on all invalid inputs.  The decode functions used
to return the number of bytes written before the problem was detected
(i.e., partial decode).  This made it difficult to distinguish this
kind of failure from success.