native_sock: Don't SIGPIPE on tx-over-closed conn
Prior to commit:

mn_sendto() on a terminated connection causes SIGPIPE to be sent to the
process (killing the process).  This is not desirable.  The signal is
not necessary; a special error code gets returned for this condition:
MN_ECONNABORTED.

After commit:

Ignore SIGPIPE at native_sock initialization time.

There are some other, less intrusive solutions (e.g., use send() rather
than write()), but they are not portable among linux and BSD.