native_sock: Don't call mn_readable during receive
Prior to commit:

native_sock_recvfrom() calls the "readable" callback when it detects a
socket error.  This is unnecessary since the socket error gets returned
by the mn_recvfrom call currently executing.  This callback can cause
confusion since the error gets reported twice (once with the callback,
once with the return code).  It can cause confusion in another way:
usually these callbacks are called from the native_sock task, but in
this case, the callback runs in the main task (or whichever task called
native_sock_recvfrom()).

Fix:

Remove the call to the "readable" callback in native_sock_recvfrom().