oic tcp4_adaptor: Don't loop if nothing to rx
Prior to commit:

When a managed TCP connection had no pending receives, the code would
enqueue another "rx" event.  As a result, the "rx" event would always be
executing or enqueued!

Fix:

If a connection has no pending receives, move on to the next connection.
Do not enqueue an "rx" event.