net/oic: Separate "stream" code from BLE transport
We use two different CoAP protocols:
    * datagram (original)
    * tcp (confusingly named)

We use "tcp" for the BLE transport and "datagram" for everything else.

The up-and-coming TCP/IPv4 transport will also use the "tcp" transport.
As a result, it will need to share some code that is currently part of
the BLE transport.

This commit moves this code out of the BLE adaptor and into a common set
of files (called stream.h and stream.c).