TCP/IPv4 adaptor for CoAP
This transport uses mn_sockets.

Unlike other transports, this one does not listen for new connections on
its own.  Instead, the application needs to populate the transport with
connections to be managed using oc_tcp4_add_conn().  There are a few too
many parameters that need to be specified when listening for TCP
connections.