Add syscfg option to select HFCLK source for NRF52 devices
This is needed to be able to run NRF52 devices without an external 32MHz crystal.

BLE will not work when set to HFINT, but everything else should.
The only restriction is that the LFCLK source cannot be LFSYNTH

In the rare case where MCU_HFCLK_SOURCE is set to HFINT and BLE is enabled, an assert(0) will be hit in nrf52_clock_hfxo_request/release