net/ip: Re-arrange lwip (#2323)
* net/ip: Re-arrange lwip

This commit changes how the mynewt lwip code is arranged in
the repository. Hopefully, this new arrangement makes it easier
to include lwip_base so other lwip implementations can use it
without having to bring the mynewt lwip code. Packages requiring
the base lwip code (and not the mynewt lwip code) can depend
on net/ip/lwip_base. Applications wanting to include the mynewt
lwip code should depend on net/ip/lwip_mn. None of the previous
functionality should change; the code was simply moved from
net/ip/src to net/ip/lwip_mn/src and a new package created.

* net/ip: Re-arrange lwip

Place a dependency to lwip_mn in the lwip driver for stm32_eth.
This is needed as the previous package in net/ip was modified
and the mynewt lwip code was moved.