Add busy I2C state recovery on STMF1/F4/L1
Sometime glitche on SDA SCL lines cause I2c to enter
busy state that never end.
All calls to read or write will fail later because of BUSY flag.
Datasheet recommens software reset in that case.
Workaround applied to transmit only, it could be on receive
but that should do.