hw/bsp/pinetime: Use spi flash for the second image
This makes it possible to double the size of the image.
To be compatible with other community firmwares the bootloader size is
increased. This causes the reboot log to shrink.
An extra SPI flash area is assigned for bootloader assets (for example
a boot logo).
The image could be increased by another 12kB, but I want to wait for
other firmwares to do that as well.