stm32: Disable WFI when OS_SYSVIEW is defined
Reading RAM or FLASH from Debugger on STM32 MCUs in SLEEP (WFI)
returns wrong values.
If makes it almost impossible to use with SystemView when application
runs at high clock speed and spend a lot of time in SLEEP mode.
SystemView is often not able to correctly read header for RTT
virtually preventing tool to detect instrumentation.

This change replaces WFI with busy loop waiting for interrupts when
OS_SYSVIEW is defined.