sys/log/cbmem: Fix return code for early break in log_cbmem_walk
The return code of succesfull log_cbmem_walk was 0, even when the loop
ended early. In case of an early loop break, we want to return the rc of
the walked function.
This is needed to make `newtmgr log show --all` work, as the rc is
passed to the client to indicate more data available.