flash_loader: Fix flash_loader.gdb to run on Windows
There were two problems with this gdb script that prevent
it usage on Windows:
1. echo command from cmd.exe don't have -n option
  shell command that was creating foo.gdb script used echo which
  did not used /usb/bin/echo.exe from msys, but Windows build in
  echo that does not have -n option
2. /tmp folder is not valid for cmd.exe redirection

To fix this problems echo is not used, instead awk that is already
present is used to output whole line
foo.gdb is not created in /tmp folder (which would fail on Windows)
instead is created in current folder for a while. It was deleted
after source command so it's not that important to have it in /tmp
folder.