da1469x: Fix DMA Multiplexer register access
Modify the crypto driver to acquire/release the
DMA channel being used. As the acquire/release
modifies a global variable and DMA_REQ_MUX_REG, which
are shared resources, these accesses need to be done
in a critical section.

The macros MCU_DMA_SET_MUX/MCU_DMA_GET_MUX are reworked
to generate the pattern for the pair of channels as documented
in the datasheet.

Configure DMA_REQ_MUX_REG within a critical section to avoid
race conditions. Use common macros provided by the DMA
driver to make changes to this register. It is to be noted that this
lock usage can be quite heavy when any given DMA channel needs to be
accessed for long periods of time and this should be addressed.

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>