hw/mcu/dialog: Fix return value from da1469x_m33_sleep
QSPIC code changes r0 register which is set in the code before to
indicate that deep sleep did not happen. This means we will always
return non-zero which means "deep sleep *did* happen" in da1469x_sleep.
It likely is not not much of a problem except we will always take
longer path when exiting from da1469x_sleep.