hw/scripts: Add GDB script to handle MTB
This adds script for GDB to easily analyze Micro Trace Buffer (MTB)
contents. A new "mtb" command is available in GDB:

- "mtb base" -> set base address for MTB, default is 0xe0043000 (CM33)
- "mtb enable" -> enables MTB (has to be already configured!)
- "mtb disable" -> disables MTB
- "mtb decode" -> decodes current MTB contents; it prints complete
                  disassembled trace along with source code if proper
                  symbol file is loaded and sources are available
- "mtb colors <on/off>" -> enables/disables colors in decoder output,
                           disabled by default and requires colorama
                           module

As a convenient shortcut, plain "mtb" command defaults to "mtb decode".
That means if MCU has MTB already configured and enabled, it's enough
to just issue that command to get valid output.