hw/mcu/dialog: Disconnect UART RX pin it stays low
When UART RX pin is shorted to ground and UART is enabled it enters
busy state.
In this state busy interrupt fires constantly it prevents non-interrupt code
to be executed and can lead to watchdog reboot.
In function hal_uart_config() pullup was added to RX pin to enable configuration
(which was not successful in busy state).
This counter measure worked fined when RX pin was floating and just happen to be
low for a while but internal pull-up resistor was enough to make is stable.
However when RX pin is shorted to ground internal pull-up will not solve the problem.

- now RX pin is not routed to UART during configuration to prevent busy state at that time.
- RX can be disconnected from UART once it enters busy state, pin is reconfigured for
GPIO with interrupt to detect end of short condition. (This functionality is optionall
and can be enabled with syscfg value MCU_UART_DICONNECT_RX_ON_BUSY).