log_fcb: optimize finding records towards the end of circular buffer
In scenarios where the interested record lies towards the end of a
large circular buffer, walking it from the beginning can take up
significant amount of time.

It is possible that the thread that walks the log is shared among
other user interfaces and until the bookmarks are built up, flash
based walks and lookups can make the system quite slow to respond.

Optimize this by beginning to walk from the flash area that is closer
to the latest one in use.

In the tests performed, the walk time was reduced by 2 orders of
magnitude (e.g 9000 ms to 50 ms).

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>