flash_map: Change sysinit stage
sysinit stage for flash map is currently 2.
It leaves very little space for additional code
that should be executed before flash_map is initialized
but after OS is created.

If some package initialization is to be executed before
flash_map, setting its stage to 1 is risky since OS is
stage 1.