sensor_creator: Split device creating and initialization
This splits existing function sensor_dev_create().
All the devices are still created in this function but code
that actually accessed them is moved to separate function
sensor_dev_init().

This is really needed in case when SPI flash is used in flash map.
Flash map is initialized in early stage. If by that time
devices that are connected to the same SPI bus are not created yet,
GPIO that is used for chip selection may be perceived as active since
it was not configured yet. In such case two devices (SPI flash and some
SPI sensor) could try to respond at the same time and SPI flash command
responses can be corrupted.