hw: drivers: hash_stm32: fix stream support
The STM32 HASH HW is configured to work on a word (32-bit) write, and
the only write that can be less than 4-bytes is the last write
(_finish()). To be compatible with streams that are non word-boundary sized,
a new pad (aka state) was added that caches writes that are less than 4
byte, so only words are ever written in an _update(), and _finish()
flushes the remaining bytes that were less than 4-bytes.

Signed-off-by: Fabio Utzig <utzig@apache.org>