hw: mcu: kinetis: fix interrupted flash programming
Disable interrupts while executing the flash commands; this avoids
hardfaults due to interrupts executing code when FCCOBx registers are
being written.

This is only protecting `FTFx_CMD_Program` which is the function that is
called from the Kinetis HAL's flash_write; other flash programming
routines might also need similar protection. I've not seen any issues
with erase operations.

A modified copy of the low-level SDK flash writing code was copied to a
standard Mynewt directory, and the original was added to ignored files,
to avoid overwriting the changes in future upgrades.

Signed-off-by: Fabio Utzig <utzig@apache.org>