hw/mcu/dialog: Latch GPIO only if COM is going to sleep
When entering sleep state of GPIO was latched regardless
whether COM would go to sleep or not.
If COM is not going to sleep register content will not be
lost and there is not need to save/restore it during MCU sleep.

Now latching GPIO is done only when last reference to COM
power domain will be released and COM domain is going to sleep.