hw: bsp: frdm-k82f: Fix flash areas and syscfgs
Update flash areas to only store the bootloader and image0 in the
internal flash, leaving other partitions in one of the QSPI flashes.

Move MCUboot mode to "move" instead of scratch, but leave a scratch
partition because FCB needs it. Use different partitions for FCB and the
FS.

Update syscfg QSPI definitions to match changes in the MCU.

Signed-off-by: Fabio Utzig <utzig@apache.org>