mcu/fe310: Fix hal timer
Hal timer was not correctly setting up compare timer resulting
in interrupts firing at wrong time.
Problem was hidden if hal timer operated at high frequencies.
For low hal_timer frequencies variable ticks is often low
and it this low value was put in CMP1 comparator while
counter was already past this value resulting in immediate
trigger of interrupt.
In CMP1 interrupt pattern was repeated.
Now CMP1 is set to value that is greater then PWMS, if it would
be grater then CMP0 (which is maximum value of counter) it is not
set at the moment of check, it will be set in CMP0 interrupt.