bsp/black_vet6: Fix ETH configuration
Some of the ETH pins were set up to port G.
There are not Port G pins on MCU on this board.
But there are pins from port B that can be used.

IRQ was hard coded to PA3 while ETH code allows
for this pin to be undefined, so now default
value is -1.

PHY address were hardcoded now they can be changed
in syscfg.