flash_map: make mfg support optional
A flash map configuration may be stored in the manufacturing meta
region, and be used instead of the one build time generated from
bsp.yml. This is currently hard-coded in the flash map code. For
users who don't use a mmr, this adds somewhere between 500 and
1000 bytes of code who serves no purpose. This PR adds a syscfg to
allow enabling/disabling the use of the MMR in the flash map thus
making the mfg dependency optional.

Signed-off-by: Fabio Utzig <utzig@apache.org>