ctype: Fix undefined behavior in ctype.h usage
Although ctype.h functions/macros like isalpha, isdigit etc.
take int argument as input behaviour is undefined if value
passed to those are not from unsigned char + EOF.
If the implementation choose to use simple macro that accesses
table indexed by argument and it was negative as it can be
when char is cast to int (as was the case in some places)
result would be unpredictable.
In most cases argument to ctype define functions was correctly
cast to unsigned (json/lwip and many other places).

This updates places that could be prone to this problem with
cast to (unsigned char).
For baselibc implementation it does not make difference but
if some other libc was chosen it could lead to potential
problems (warnings at build time or unpredictable behavior).
Line from code build against non-baselibc implementation of ctype.h
error: array subscript has type 'char' [-Werror=char-subscripts]