mcu/da1469x: Move .init_array sections to flash
content of .init_array* sections is meant to be executed only
once during system start.
For no apparent reason those arrays were put in data section.
This would results in wast of RAM for C++ code built if it was
not for the fact that KEEP keyword was missing so such sections
would not result in final build (RAM nor FLASH) because by
compiler design  content of those section is not referenced
explicitly and relays on linker using them without apparent
reference.

Sections are now moved to .text and used KEEP keyword to make
sure constructors will be executed.