da1469x/i2c: Correct CON_REG settings
For register I2C_CON_REG datasheet description of
bit I2C_MASTER_MODE:
This bit controls whether the controller master is enabled.
 0= master disabled
 1= master enabled
Software should ensure that if this bit is written with
'1' then bit 6 (SLAVE_DISABLE) should also be written with a '1'.

It's not clear what will happened otherwise, so far it seemed to
work without this but to be on the safe side.