bus/i2c_nrf5x: Assert on not supported usage
Current code return error code when read was called
with BUS_F_NOSTOP flag set.

It seems better to just assert at line where the user can
see why this not working instead of returning error code
where user will have to figure out why I2C lines are dead.
This kind of problem is not something that can be fixed
in runtime by just repeating read it will always fail
until the code is modified not to use such flag.

This change should alert the user that this kind of desired
behavior will not work on this platform.