hal/flash: Relax flash id number restriction
So far flash ids needed to start from 0 to some number.
It was not possible to have only some numbers.
If BSP specified flash areas that belonged to 3 devices.
Device ID would be 0,1,2.
If bootloader (or some application configuration) was not
to touch some of the devices hal_bsp_flash_dev() function
would just return NULL to end flash device enumeration.
If bsp.yml contained flash areas belonging to 3 flash devices,
application (bootloader) could have access to 0, or 0,1 or 0,1,2
devices. It could not access only 0 and 2.

This change introduces HAL_FLASH_MAX_DEVICE_COUNT that when specified
allows to not have driver present for some of the flash devices.

Default value (0) keeps old behavior.