hw/bus/spi_hal: Fix bus_spi_write_read
Function bus_spi_write_read() was calling hal_spi_txrx()
with NULL for txbuf for transactions involving more then 16 bytes.
NULL pointer is not valid for this function.
When code was build with SPI_HAL_USE_NOBLOCK enabled, zeroed receive buffer
was already used as argument for transmit buffer.
Same approach is now used when SPI_HAL_USE_NOBLOCK is not set.

For data reads 0xFF is used for MOSI line just to be consistent.
Some places used 0 some had 0xFF now 0xFF is used as it was
originally.